+++ Source/Urho3D/Container/HashSet.h
@@ -41,24 +41,26 @@ public:
         Node()
         {
         }
-        
+
         /// Construct with key.
         Node(const T& key) :
             key_(key)
         {
         }
-        
+
         /// Key.
         T key_;
-        
+
         /// Return next node.
         Node* Next() const { return static_cast<Node*>(next_); }
+
         /// Return previous node.
         Node* Prev() const { return static_cast<Node*>(prev_); }
+
         /// Return next node in the bucket.
         Node* Down() const { return static_cast<Node*>(down_); }
     };
-    
+
     /// Hash set node iterator.
     struct Iterator : public HashIteratorBase
     {
@@ -66,28 +68,50 @@ public:
         Iterator()
         {
         }
-        
+
         /// Construct with a node pointer.
         Iterator(Node* ptr) :
             HashIteratorBase(ptr)
         {
         }
-        
+
         /// Preincrement the pointer.
-        Iterator& operator ++ () { GotoNext(); return *this; }
+        Iterator& operator ++()
+        {
+            GotoNext();
+            return *this;
+        }
+
         /// Postincrement the pointer.
-        Iterator operator ++ (int) { Iterator it = *this; GotoNext(); return it; }
+        Iterator operator ++(int)
+        {
+            Iterator it = *this;
+            GotoNext();
+            return it;
+        }
+
         /// Predecrement the pointer.
-        Iterator& operator -- () { GotoPrev(); return *this; }
+        Iterator& operator --()
+        {
+            GotoPrev();
+            return *this;
+        }
+
         /// Postdecrement the pointer.
-        Iterator operator -- (int) { Iterator it = *this; GotoPrev(); return it; }
-        
+        Iterator operator --(int)
+        {
+            Iterator it = *this;
+            GotoPrev();
+            return it;
+        }
+
         /// Point to the key.
-        const T* operator -> () const { return &(static_cast<Node*>(ptr_))->key_; }
+        const T* operator ->() const { return &(static_cast<Node*>(ptr_))->key_; }
+
         /// Dereference the key.
-        const T& operator * () const { return (static_cast<Node*>(ptr_))->key_; }
+        const T& operator *() const { return (static_cast<Node*>(ptr_))->key_; }
     };
-    
+
     /// Hash set node const iterator.
     struct ConstIterator : public HashIteratorBase
     {
@@ -95,53 +119,80 @@ public:
         ConstIterator()
         {
         }
-        
+
         /// Construct with a node pointer.
         ConstIterator(Node* ptr) :
             HashIteratorBase(ptr)
         {
         }
-        
+
         /// Construct from a non-const iterator.
         ConstIterator(const Iterator& rhs) :
             HashIteratorBase(rhs.ptr_)
         {
         }
-        
+
         /// Assign from a non-const iterator.
-        ConstIterator& operator = (const Iterator& rhs) { ptr_ = rhs.ptr_; return *this; }
+        ConstIterator& operator =(const Iterator& rhs)
+        {
+            ptr_ = rhs.ptr_;
+            return *this;
+        }
+
         /// Preincrement the pointer.
-        ConstIterator& operator ++ () { GotoNext(); return *this; }
+        ConstIterator& operator ++()
+        {
+            GotoNext();
+            return *this;
+        }
+
         /// Postincrement the pointer.
-        ConstIterator operator ++ (int) { ConstIterator it = *this; GotoNext(); return it; }
+        ConstIterator operator ++(int)
+        {
+            ConstIterator it = *this;
+            GotoNext();
+            return it;
+        }
+
         /// Predecrement the pointer.
-        ConstIterator& operator -- () { GotoPrev(); return *this; }
+        ConstIterator& operator --()
+        {
+            GotoPrev();
+            return *this;
+        }
+
         /// Postdecrement the pointer.
-        ConstIterator operator -- (int) { ConstIterator it = *this; GotoPrev(); return it; }
-        
+        ConstIterator operator --(int)
+        {
+            ConstIterator it = *this;
+            GotoPrev();
+            return it;
+        }
+
         /// Point to the key.
-        const T* operator -> () const { return &(static_cast<Node*>(ptr_))->key_; }
+        const T* operator ->() const { return &(static_cast<Node*>(ptr_))->key_; }
+
         /// Dereference the key.
-        const T& operator * () const { return (static_cast<Node*>(ptr_))->key_; }
+        const T& operator *() const { return (static_cast<Node*>(ptr_))->key_; }
     };
-    
+
     /// Construct empty.
     HashSet()
     {
         // Reserve the tail node
-        allocator_ = AllocatorInitialize(sizeof(Node));
+        allocator_ = AllocatorInitialize((unsigned)sizeof(Node));
         head_ = tail_ = ReserveNode();
     }
-    
+
     /// Construct from another hash set.
     HashSet(const HashSet<T>& set)
     {
         // Reserve the tail node + initial capacity according to the set's size
-        allocator_ = AllocatorInitialize(sizeof(Node), set.Size() + 1);
+        allocator_ = AllocatorInitialize((unsigned)sizeof(Node), set.Size() + 1);
         head_ = tail_ = ReserveNode();
         *this = set;
     }
-    
+
     /// Destruct.
     ~HashSet()
     {
@@ -150,35 +201,35 @@ public:
         AllocatorUninitialize(allocator_);
         delete[] ptrs_;
     }
-    
+
     /// Assign a hash set.
-    HashSet& operator = (const HashSet<T>& rhs)
+    HashSet& operator =(const HashSet<T>& rhs)
     {
         Clear();
         Insert(rhs);
         return *this;
     }
-    
+
     /// Add-assign a value.
-    HashSet& operator += (const T& rhs)
+    HashSet& operator +=(const T& rhs)
     {
         Insert(rhs);
         return *this;
     }
-    
+
     /// Add-assign a hash set.
-    HashSet& operator += (const HashSet<T>& rhs)
+    HashSet& operator +=(const HashSet<T>& rhs)
     {
         Insert(rhs);
         return *this;
     }
-    
+
     /// Test for equality with another hash set.
-    bool operator == (const HashSet<T>& rhs) const
+    bool operator ==(const HashSet<T>& rhs) const
     {
         if (rhs.Size() != Size())
             return false;
-        
+
         ConstIterator it = Begin();
         while (it != End())
         {
@@ -186,16 +237,16 @@ public:
                 return false;
             ++it;
         }
-        
+
         return true;
     }
-    
+
     /// Test for inequality with another hash set.
-    bool operator != (const HashSet<T>& rhs) const
+    bool operator !=(const HashSet<T>& rhs) const
     {
         if (rhs.Size() != Size())
             return true;
-        
+
         ConstIterator it = Begin();
         while (it != End())
         {
@@ -203,10 +254,10 @@ public:
                 return true;
             ++it;
         }
-        
+
         return false;
     }
-    
+
     /// Insert a key. Return an iterator to it.
     Iterator Insert(const T& key)
     {
@@ -216,27 +267,27 @@ public:
             AllocateBuckets(Size(), MIN_BUCKETS);
             Rehash();
         }
-        
+
         unsigned hashKey = Hash(key);
-        
+
         Node* existing = FindNode(key, hashKey);
         if (existing)
             return Iterator(existing);
-        
+
         Node* newNode = InsertNode(Tail(), key);
         newNode->down_ = Ptrs()[hashKey];
         Ptrs()[hashKey] = newNode;
-        
+
         // Rehash if the maximum load factor has been exceeded
         if (Size() > NumBuckets() * MAX_LOAD_FACTOR)
         {
             AllocateBuckets(Size(), NumBuckets() << 1);
             Rehash();
         }
-        
+
         return Iterator(newNode);
     }
-    
+
     /// Insert a set.
     void Insert(const HashSet<T>& set)
     {
@@ -245,46 +296,46 @@ public:
         while (it != end)
             Insert(*it++);
     }
-    
+
     /// Insert a key by iterator. Return iterator to the value.
     Iterator Insert(const ConstIterator& it)
     {
         return Iterator(InsertNode(*it));
     }
-    
+
     /// Erase a key. Return true if was found.
     bool Erase(const T& key)
     {
         if (!ptrs_)
             return false;
-        
+
         unsigned hashKey = Hash(key);
-        
+
         Node* previous;
         Node* node = FindNode(key, hashKey, previous);
         if (!node)
             return false;
-        
+
         if (previous)
             previous->down_ = node->down_;
         else
             Ptrs()[hashKey] = node->down_;
-        
+
         EraseNode(node);
         return true;
     }
-    
+
     /// Erase a key by iterator. Return iterator to the next key.
     Iterator Erase(const Iterator& it)
     {
         if (!ptrs_ || !it.ptr_)
             return End();
-        
+
         Node* node = static_cast<Node*>(it.ptr_);
         Node* next = node->Next();
-        
+
         unsigned hashKey = Hash(node->key_);
-        
+
         Node* previous = 0;
         Node* current = static_cast<Node*>(Ptrs()[hashKey]);
         while (current && current != node)
@@ -292,54 +343,54 @@ public:
             previous = current;
             current = current->Down();
         }
-        
+
         assert(current == node);
-        
+
         if (previous)
             previous->down_ = node->down_;
         else
             Ptrs()[hashKey] = node->down_;
-        
+
         EraseNode(node);
         return Iterator(next);
     }
-    
+
     /// Clear the set.
     void Clear()
     {
         if (Size())
         {
-            for (Iterator i = Begin(); i != End(); )
+            for (Iterator i = Begin(); i != End();)
             {
                 FreeNode(static_cast<Node*>(i++.ptr_));
                 i.ptr_->prev_ = 0;
             }
-            
+
             head_ = tail_;
             SetSize(0);
         }
-        
+
         ResetPtrs();
     }
-    
+
     /// Sort keys. After sorting the set can be iterated in order until new elements are inserted.
     void Sort()
     {
         unsigned numKeys = Size();
         if (!numKeys)
             return;
-        
-        Node** ptrs = new Node*[numKeys];
+
+        Node** ptrs = new Node* [numKeys];
         Node* ptr = Head();
-        
+
         for (unsigned i = 0; i < numKeys; ++i)
         {
             ptrs[i] = ptr;
             ptr = ptr->Next();
         }
-        
+
         Urho3D::Sort(RandomAccessIterator<Node*>(ptrs), RandomAccessIterator<Node*>(ptrs + numKeys), CompareNodes);
-        
+
         head_ = ptrs[0];
         ptrs[0]->prev_ = 0;
         for (unsigned i = 1; i < numKeys; ++i)
@@ -349,10 +400,10 @@ public:
         }
         ptrs[numKeys - 1]->next_ = tail_;
         tail_->prev_ = ptrs[numKeys - 1];
-        
+
         delete[] ptrs;
     }
-    
+
     /// Rehash to a specific bucket count, which must be a power of two. Return true if successful.
     bool Rehash(unsigned numBuckets)
     {
@@ -360,25 +411,25 @@ public:
             return true;
         if (!numBuckets || numBuckets < Size() / MAX_LOAD_FACTOR)
             return false;
-        
+
         // Check for being power of two
         unsigned check = numBuckets;
         while (!(check & 1))
             check >>= 1;
         if (check != 1)
             return false;
-        
+
         AllocateBuckets(Size(), numBuckets);
         Rehash();
         return true;
     }
-    
+
     /// Return iterator to the key, or end iterator if not found.
     Iterator Find(const T& key)
     {
         if (!ptrs_)
             return End();
-        
+
         unsigned hashKey = Hash(key);
         Node* node = FindNode(key, hashKey);
         if (node)
@@ -386,13 +437,13 @@ public:
         else
             return End();
     }
-    
+
     /// Return const iterator to the key, or end iterator if not found.
     ConstIterator Find(const T& key) const
     {
         if (!ptrs_)
             return End();
-        
+
         unsigned hashKey = Hash(key);
         Node* node = FindNode(key, hashKey);
         if (node)
@@ -400,36 +451,42 @@ public:
         else
             return End();
     }
-    
+
     /// Return whether contains a key.
     bool Contains(const T& key) const
     {
         if (!ptrs_)
             return false;
-        
+
         unsigned hashKey = Hash(key);
         return FindNode(key, hashKey) != 0;
     }
-    
+
     /// Return iterator to the beginning.
     Iterator Begin() { return Iterator(Head()); }
+
     /// Return iterator to the beginning.
     ConstIterator Begin() const { return ConstIterator(Head()); }
+
     /// Return iterator to the end.
     Iterator End() { return Iterator(Tail()); }
+
     /// Return iterator to the end.
     ConstIterator End() const { return ConstIterator(Tail()); }
+
     /// Return first key.
     const T& Front() const { return *Begin(); }
+
     /// Return last key.
     const T& Back() const { return *(--End()); }
-    
+
 private:
     /// Return the head node.
     Node* Head() const { return static_cast<Node*>(head_); }
+
     /// Return the tail node.
     Node* Tail() const { return static_cast<Node*>(tail_); }
-    
+
     /// Find a node from the buckets. Do not call if the buckets have not been allocated.
     Node* FindNode(const T& key, unsigned hashKey) const
     {
@@ -440,15 +497,15 @@ private:
                 return node;
             node = node->Down();
         }
-        
+
         return 0;
     }
-    
+
     /// Find a node and the previous node from the buckets. Do not call if the buckets have not been allocated.
     Node* FindNode(const T& key, unsigned hashKey, Node*& previous) const
     {
         previous = 0;
-        
+
         Node* node = static_cast<Node*>(Ptrs()[hashKey]);
         while (node)
         {
@@ -457,16 +514,16 @@ private:
             previous = node;
             node = node->Down();
         }
-        
+
         return 0;
     }
-    
+
     /// Insert a node into the list. Return the new node.
     Node* InsertNode(Node* dest, const T& key)
     {
         if (!dest)
             return 0;
-        
+
         Node* newNode = ReserveNode(key);
         Node* prev = dest->Prev();
         newNode->next_ = dest;
@@ -474,39 +531,39 @@ private:
         if (prev)
             prev->next_ = newNode;
         dest->prev_ = newNode;
-        
+
         // Reassign the head node if necessary
         if (dest == Head())
             head_ = newNode;
-        
+
         SetSize(Size() + 1);
-        
+
         return newNode;
     }
-    
+
     /// Erase a node from the list. Return pointer to the next element, or to the end if could not erase.
     Node* EraseNode(Node* node)
     {
         // The tail node can not be removed
         if (!node || node == tail_)
             return Tail();
-        
+
         Node* prev = node->Prev();
         Node* next = node->Next();
         if (prev)
             prev->next_ = next;
         next->prev_ = prev;
-        
+
         // Reassign the head node if necessary
         if (node == Head())
             head_ = next;
-        
+
         FreeNode(node);
         SetSize(Size() - 1);
-        
+
         return next;
     }
-    
+
     /// Reserve a node.
     Node* ReserveNode()
     {
@@ -514,7 +571,7 @@ private:
         new(newNode) Node();
         return newNode;
     }
-    
+
     /// Reserve a node with specified key.
     Node* ReserveNode(const T& key)
     {
@@ -522,14 +579,14 @@ private:
         new(newNode) Node(key);
         return newNode;
     }
-    
+
     /// Free a node.
     void FreeNode(Node* node)
     {
         (node)->~Node();
         AllocatorFree(allocator_, node);
     }
-    
+
     /// Rehash the buckets.
     void Rehash()
     {
@@ -541,7 +598,7 @@ private:
             Ptrs()[hashKey] = node;
         }
     }
-    
+
     /// Compare two nodes.
     static bool CompareNodes(Node*& lhs, Node*& rhs) { return lhs->key_ < rhs->key_; }
 
@@ -555,8 +612,11 @@ namespace std
 {
 
 template <class T> typename Urho3D::HashSet<T>::ConstIterator begin(const Urho3D::HashSet<T>& v) { return v.Begin(); }
+
 template <class T> typename Urho3D::HashSet<T>::ConstIterator end(const Urho3D::HashSet<T>& v) { return v.End(); }
+
 template <class T> typename Urho3D::HashSet<T>::Iterator begin(Urho3D::HashSet<T>& v) { return v.Begin(); }
+
 template <class T> typename Urho3D::HashSet<T>::Iterator end(Urho3D::HashSet<T>& v) { return v.End(); }
 
 }
